Refactor property prefixes into `GradlePropertiesLoader`
