+++ subprojects/core/src/test/groovy/org/gradle/initialization/DefaultProjectPropertiesLoaderTest.groovy
@@ -22,8 +22,8 @@ import org.gradle.initialization.properties.ProjectPropertiesLoader
 import spock.lang.Specification
 
 import static java.util.Collections.emptyMap
-import static org.gradle.initialization.IGradlePropertiesLoader.ENV_PROJECT_PROPERTIES_PREFIX
-import static org.gradle.initialization.IGradlePropertiesLoader.SYSTEM_PROJECT_PROPERTIES_PREFIX
+import static org.gradle.initialization.properties.GradlePropertiesLoader.ENV_PROJECT_PROPERTIES_PREFIX
+import static org.gradle.initialization.properties.GradlePropertiesLoader.SYSTEM_PROJECT_PROPERTIES_PREFIX
 
 class DefaultProjectPropertiesLoaderTest extends Specification {
